add delete alert box

+++ admin_panel/src/app/views/pages/core/admin/content-authorizer/view-authorizer/view-authorizers.component.ts
@@ -98,16 +98,19 @@ export class ViewAuthorizersComponent implements OnInit {
     /**
      * delete authorizer
      */
-    deleteAuthorizer(deleteId){
-        this.authorizerService.deleteAuthorizer(
-            deleteId
-        ).subscribe(
-            success => {
-                this.authorizerDeletingStatus = success.success;
-                this.error = success.error;
-                this.getAuthorizerDetails();
-                this.hideAlert();
-            }
-        );
+    deleteAuthorizer(deleteId, name){
+        if(confirm("Are you sure to delete ' "+ name+" ' ?")){
+            this.authorizerService.deleteAuthorizer(
+                deleteId
+            ).subscribe(
+                success => {
+                    this.authorizerDeletingStatus = success.success;
+                    this.error = success.error;
+                    this.getAuthorizerDetails();
+                    this.hideAlert();
+                }
+                );
+        }
+        
     }
 }
